Move poll DAL
